PLAT-15858 Update documentation and remove language detection code (#248)
* Normalizes CodeQL tool driver name matching

Ensures that the CodeQL tool driver name is correctly overwritten to 'GitHub CodeQL' regardless of case or surrounding whitespace.

Updates the logic to trim and lowercase the tool driver name before comparing it to "codeql", handling variations in casing and spacing.
Also adds unit tests to verify the fix works for different cases.

* Bump dependency versions

* More updates

* Remove unused github token

* Minor updates to README

---------

Co-authored-by: Colin Wong <cwong@securecodewarrior.com>